+++ lesson3_vera/my_map.rb
@@ -1,13 +1,5 @@
 # frozen_string_literal: true
 
-class Array
-  def my_map
-    clone = []
-    each do |x|
-      clone.push yield x
-    end
-    clone
-  end
-end
+require_relative 'lib/array'
 arr = [1, 2, 3, 4, 5]
-print(arr.my_map { |n| n * n })
+p(arr.my_map { |n| n * n })
