+++ lesson3_vera/my_reduce.rb
@@ -1,11 +1,5 @@
 # frozen_string_literal: true
 
-class Array
-  def my_reduce(val = first)
-    map do |el|
-      val = yield(val, el)
-    end
-    val
-  end
-end
-puts([1, 2, 3, 4, 5].my_reduce { |a, b| a * b })
+require_relative 'lib/array'
+arr = [1, 2, 3, 4, 5]
+puts(arr.my_reduce { |a, b| a * b })
