+++ lesson3_vera/my_select.rb
@@ -1,13 +1,6 @@
 # frozen_string_literal: true
 
-class Array
-  def my_select
-    clone = []
-    each do |el|
-      clone.push(el) if yield el
-    end
-    clone
-  end
-end
+require_relative 'lib/array'
+
 arr = [1, 2, 3, 4, 5]
-p arr.my_select(&:odd?)
+p arr.my_select(&:even?)
