+++ lesson3_vera/weekends.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'date'
+# require_relative 'lib/date'
 
 class Date
   def weekends
